Refactor parts of Pico integration
