+++ inc/integrations/namespace.php
@@ -55,6 +55,7 @@ function auto_register_integrations() {
 function auto_register_plugin_integrations() {
 	$integrations = [
 		'jwt_auth' => __NAMESPACE__ . '\JWT_Auth',
+		'pico'     => __NAMESPACE__ . '\Pico',
 	];
 
 	// Allow new integrations to be added via a filter.
